+++ .github/CODEOWNERS
@@ -8,3 +8,4 @@
 
 # General ownership: Next Build Team (ap-team) & Platform FE COP
 *       @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
+src/products/eventListing/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
