+++ src/data/queries/index.ts
@@ -13,7 +13,7 @@ import * as CollapsiblePanelItem from './collapsiblePanelItem'
 import * as ContactInfo from './contactInfo'
 import * as EmailContact from './emailContact'
 import * as Event from './event'
-import * as EventListing from './eventListing'
+import * as EventListing from '@/products/eventListing/query'
 import * as EventTeaser from './eventTeaser'
 import * as ExpandableText from './expandableText'
 import * as FeaturedContent from './featuredContent'
