+++ src/products/eventListing/query.test.ts
@@ -4,10 +4,10 @@
 
 import { NodeEvent, NodeEventListing } from '@/types/drupal/node'
 import { queries } from '@/data/queries'
-import { mockResponse } from '@/mocks/eventListing.mock.js'
+import { mockResponse } from '@/products/eventListing/mock.js'
 import mockEventData from '@/mocks/event.mock.json'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
-import { params } from '../eventListing'
+import { params } from './query'
 
 const EventListingMock: NodeEventListing = mockResponse
 const EventMock: NodeEvent[] = [mockEventData]
