+++ src/products/eventListing/query.ts
@@ -1,9 +1,9 @@
 import { QueryData, QueryFormatter, QueryParams } from 'next-drupal-query'
-import { queries } from '.'
+import { queries } from '@/data/queries'
 import { DrupalJsonApiParams } from 'drupal-jsonapi-params'
 import { NodeEvent, NodeEventListing } from '@/types/drupal/node'
 import { Menu } from '@/types/drupal/menu'
-import { EventListing } from '@/types/formatted/eventListing'
+import { EventListing } from '@/products/eventListing/formatted-type'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
 import { PAGE_SIZES } from '@/lib/constants/pageSizes'
 import { ExpandedStaticPropsContext } from '@/lib/drupal/staticProps'
